+++ pkg/stave/templates/initial_stavefile.go
@@ -1,5 +1,4 @@
 //go:build stave
-// +build stave
 
 package main
 
@@ -8,37 +7,47 @@ import (
 	"os"
 	"os/exec"
 
+	"github.com/charmbracelet/log"
 	"github.com/yaklabco/stave/pkg/st" // st contains helpful utility functions, like Deps
+	"github.com/yaklabco/stave/pkg/stave/prettylog"
 )
+
+func init() { //nolint:gochecknoinits // This is a stavefile, we can have an init().
+	logHandler := prettylog.SetupPrettyLogger(os.Stdout)
+	if st.Debug() {
+		logHandler.SetLevel(log.DebugLevel)
+	}
+}
 
 // Default target to run when none is specified
 // If not set, running stave will list available targets
 // var Default = Build
 
-// A build step that requires additional params, or platform specific steps for example
+// A build step that requires additional params, or platform specific steps for example.
 func Build() error {
 	st.Deps(InstallDeps)
-	fmt.Println("Building...")
+	_, _ = fmt.Println("Building...")
 	cmd := exec.Command("go", "build", "-o", "MyApp", ".")
 	return cmd.Run()
 }
 
-// A custom install step if you need your bin someplace other than go/bin
+// A custom install step if you need your bin someplace other than go/bin.
 func Install() error {
 	st.Deps(Build)
-	fmt.Println("Installing...")
+	_, _ = fmt.Println("Installing...")
 	return os.Rename("./MyApp", "/usr/bin/MyApp")
 }
 
 // Manage your deps, or running package managers.
 func InstallDeps() error {
-	fmt.Println("Installing Deps...")
+	_, _ = fmt.Println("Installing Deps...")
 	cmd := exec.Command("go", "get", "github.com/stretchr/piglatin")
 	return cmd.Run()
 }
 
-// Clean up after yourself
-func Clean() {
-	fmt.Println("Cleaning...")
-	os.RemoveAll("MyApp")
+// Clean up after yourself.
+func Clean() error {
+	_, _ = fmt.Println("Cleaning...")
+
+	return os.RemoveAll("MyApp")
 }