+++ pkg/stave/testdata/list/command.go
@@ -1,5 +1,4 @@
 //go:build stave
-// +build stave
 
 // This is a comment on the package which should get turned into output with the
 // list of targets.