+++ pkg/stave/testdata/setworkdir/stavefile.go
@@ -1,16 +1,15 @@
 //go:build stave
-// +build stave
 
 package main
 
 import (
 	"fmt"
-	"io/ioutil"
+	"os"
 	"strings"
 )
 
 func TestWorkingDir() error {
-	files, err := ioutil.ReadDir(".")
+	files, err := os.ReadDir(".")
 	if err != nil {
 		return err
 	}